[conf] moved configuration code into its own package
