+++ setup.cfg
@@ -1,4 +1,4 @@
 [nosetests]
 all-modules=1
 with-coverage=1
-cover-package=libreantdb, webant, presets, archivant, utils, cli
+cover-package=libreantdb, webant, presets, archivant, utils, cli, conf
