+++ setup.py
@@ -94,7 +94,14 @@ conf = dict(
         author_email='insomnialab@hacari.org',
         url='https://github.com/insomnia-lab/libreant',
         license='AGPL',
-        packages=['libreantdb', 'webant', 'webant.api', 'presets', 'archivant', 'utils', 'cli'],
+        packages=['libreantdb',
+                  'webant',
+                  'webant.api',
+                  'presets',
+                  'archivant',
+                  'utils',
+                  'cli',
+                  'conf'],
         install_requires=[
           'gevent',
           'elasticsearch',
